+++ Mkfiles/msvc.mak
@@ -86,53 +86,21 @@ NDISASM = disasm\ndisasm.obj
 PROGOBJ = $(NASM) $(NDISASM)
 PROGS   = nasm$(X) ndisasm$(X)
 
-# Files dependent on warnings.dat
-WARNOBJ   = asm\warnings.obj
-WARNFILES = asm\warnings_c.h include\warnings.h doc\warnings.src
+# Objects for the local copy of zlib. The variable ZLIB is set to
+# $(ZLIBOBJ) if the internal version of zlib should be used.
+ZLIBOBJ = \
+	zlib\adler32.obj \
+	zlib\crc32.obj \
+	zlib\infback.obj \
+	zlib\inffast.obj \
+	zlib\inflate.obj \
+	zlib\inftrees.obj \
+	zlib\zutil.obj
 
-OUTPUTOBJ = \
-	output\outform.obj output\outlib.obj \
-	output\nulldbg.obj output\nullout.obj \
-	output\outbin.obj output\outaout.obj output\outcoff.obj \
-	output\outelf.obj \
-	output\outobj.obj output\outas86.obj \
-	output\outdbg.obj output\outieee.obj output\outmacho.obj \
-	output\codeview.obj
-
-# The source files for these objects are scanned for warnings
-LIBOBJ_W = \
-	nasmlib\readnum.obj \
-	\
-	asm\error.obj \
-	asm\floats.obj \
-	asm\directiv.obj \
-	asm\pragma.obj \
-	asm\assemble.obj asm\labels.obj asm\parser.obj \
-	asm\preproc.obj asm\quote.obj \
-	asm\listing.obj asm\eval.obj asm\exprlib.obj asm\exprdump.obj \
-	asm\stdscan.obj \
-	asm\getbool.obj \
-	asm\strfunc.obj \
-	asm\segalloc.obj \
-	asm\rdstrnum.obj \
-	asm\srcfile.obj \
-	\
-	$(OUTPUTOBJ)
-
-# The source files for these objects are NOT scanned for warnings;
-# normally this will include all generated files.
-# It is entirely possible that it may be necessary to move some of these
-# files to LIBOBJ_W, notably $(OUTPUTOBJ)
-LIBOBJ_NW = \
+# Common library objects
+LIBOBJ_COM = \
 	stdlib\snprintf.obj stdlib\vsnprintf.obj stdlib\strlcpy.obj \
 	stdlib\strnlen.obj stdlib\strrchrnul.obj \
-	\
-	asm\directbl.obj \
-	asm\pptok.obj \
-	asm\tokhash.obj \
-	asm\uncompress.obj \
-	\
-	macros\macros.obj \
 	\
 	nasmlib\ver.obj \
 	nasmlib\alloc.obj nasmlib\asprintf.obj \
@@ -147,35 +115,63 @@ LIBOBJ_NW = \
 	nasmlib\raa.obj nasmlib\saa.obj \
 	nasmlib\strlist.obj \
 	nasmlib\perfhash.obj nasmlib\badenum.obj \
+	nasmlib\readnum.obj \
 	\
-	common\common.obj \
+	common\common.obj common\errstubs.obj \
 	\
 	x86\insnsa.obj x86\insnsb.obj x86\insnsn.obj \
 	x86\regs.obj x86\regvals.obj x86\regflags.obj \
 	x86\iflag.obj \
 	\
-	$(WARNOBJ)
+	$(ZLIB)
+
+# Files dependent on warnings.dat
+WARNOBJ   = asm\warnings.obj
+WARNFILES = asm\warnings_c.h include\warnings.h doc\warnings.src
+
+OUTPUTOBJ = \
+	output\outform.obj output\outlib.obj \
+	output\nulldbg.obj output\nullout.obj \
+	output\outbin.obj output\outaout.obj output\outcoff.obj \
+	output\outelf.obj \
+	output\outobj.obj output\outas86.obj \
+	output\outdbg.obj output\outieee.obj output\outmacho.obj \
+	output\codeview.obj
+
+# Assembler-only library objects
+LIBOBJ_ASM = \
+	asm\error.obj \
+	asm\floats.obj \
+	asm\directiv.obj \
+	asm\pragma.obj \
+	asm\assemble.obj asm\labels.obj asm\parser.obj \
+	asm\preproc.obj asm\quote.obj \
+	asm\listing.obj asm\eval.obj asm\exprlib.obj asm\exprdump.obj \
+	asm\stdscan.obj \
+	asm\getbool.obj \
+	asm\strfunc.obj \
+	asm\segalloc.obj \
+	asm\rdstrnum.obj \
+	asm\srcfile.obj \
+	asm\directbl.obj \
+	asm\pptok.obj \
+	asm\tokhash.obj \
+	asm\uncompress.obj \
+	\
+	macros\macros.obj \
+	\
+	$(WARNOBJ) \
+	$(OUTPUTOBJ)
 
 # Objects which are only used for the disassembler
 LIBOBJ_DIS = \
 	disasm\disasm.obj disasm\sync.obj disasm\prefix.obj \
+	disasm\diserror.obj \
 	\
 	x86\insnsd.obj x86\regdis.obj
 
-# Objects for the local copy of zlib. The variable ZLIB is set to
-# $(ZLIBOBJ) if the internal version of zlib should be used.
-ZLIBOBJ = \
-	zlib\adler32.obj \
-	zlib\crc32.obj \
-	zlib\infback.obj \
-	zlib\inffast.obj \
-	zlib\inflate.obj \
-	zlib\inftrees.obj \
-	zlib\zutil.obj
-
-LIBOBJ    = $(LIBOBJ_W) $(LIBOBJ_NW) $(ZLIB)
-ALLOBJ_W  = $(NASM) $(LIBOBJ_W)
-ALLOBJ    = $(PROGOBJ) $(LIBOBJ) $(LIBOBJ_DIS)
+LIBOBJ    = $(LIBOBJ_COM) $(LIBOBJ_ASM) $(LIBOBJ_DIS)
+ALLOBJ    = $(PROGOBJ) $(LIBOBJ)
 SUBDIRS  = stdlib nasmlib include config output asm disasm x86 \
 	   common zlib macros misc
 XSUBDIRS = nsis win test doc editors
@@ -186,22 +182,26 @@ EDITORS  = editors\nasmtok.el editors\nasmtok.json
 #-- End File Lists --#
 
 NASMLIB = libnasm.$(A)
-NDISLIB = libndis.$(A)
+ASMLIB  = libasm.$(A)
+DISLIB  = libdis.$(A)
 
 all: nasm$(X) ndisasm$(X)
 
-nasm$(X): $(NASM) $(MANIFEST) $(NASMLIB)
-	$(CC) /Fe:$@ $(ALL_CFLAGS) $(NASM) $(NASMLIB) $(LIBS) \
+nasm$(X): $(NASM) $(MANIFEST) $(ASMLIB) $(NASMLIB)
+	$(CC) /Fe:$@ $(ALL_CFLAGS) $(NASM) $(ASMLIB) $(NASMLIB) $(LIBS) \
 		$(ALL_LDFLAGS)
 
-ndisasm$(X): $(NDISASM) $(MANIFEST) $(NDISLIB) $(NASMLIB)
-	$(CC) /Fe:$@ $(ALL_CFLAGS) $(NDISASM) $(NDISLIB) $(NASMLIB) $(LIBS) \
+ndisasm$(X): $(NDISASM) $(MANIFEST) $(DISLIB) $(NASMLIB)
+	$(CC) /Fe:$@ $(ALL_CFLAGS) $(NDISASM) $(DISLIB) $(NASMLIB) $(LIBS) \
 		$(ALL_LDFLAGS)
 
-$(NASMLIB): $(LIBOBJ)
+$(NASMLIB): $(LIBOBJ_COM)
 	$(AR) $(ARFLAGS) /out:$@ $**
 
-$(NDISLIB): $(LIBOBJ_DIS)
+$(ASMLIB): $(LIBOBJ_ASM)
+	$(AR) $(ARFLAGS) /out:$@ $**
+
+$(DISLIB): $(LIBOBJ_DIS)
 	$(AR) $(ARFLAGS) /out:$@ $**
 
 # These are specific to certain Makefile syntaxes...
